requirements.txt
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
passlib==1.7.4
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
typing==3.7.4.3
typing_extensions==4.4.0
uvicorn==0.20.0
